begin of setup for implementing more complex indices. removed the shape wrapper type and keeping only the strict type for now until a decision is made about what types of indices will be supported, including stacked RLE elements i.e. soil layers
