+++ python/source/soillib.cpp
@@ -9,6 +9,7 @@
 // Bind Function Declarations
 
 namespace nb = nanobind;
+void bind_index(nb::module_& module);
 void bind_io(nb::module_& module);
 void bind_util(nb::module_& module);
 void bind_layer(nb::module_& module);
@@ -23,7 +24,8 @@ nb::set_leak_warnings(false);
 
 module.doc() = "Soillib Python Module";
 
-bind_io (module);
+bind_index(module);
+bind_io(module);
 bind_util(module);
 bind_layer(module);
 bind_matrix(module);
