+++ source/index/quad.hpp
@@ -1,0 +1,12 @@
+#ifndef SOILLIB_INDEX_QUAD
+#define SOILLIB_INDEX_QUAD
+
+namespace soil {
+
+//! basic quadtree structured index\
+//! \todo implement
+struct quad{};
+
+} // end of namespace soil
+
+#endif
